requirements.txt
aiohttp==3.8.4
asyncio==3.4.3
datasets==2.12.0
ipykernel==6.22.0
Jinja2==3.1.2
joblib==1.2.0
loguru==0.6.0
matplotlib==3.7.1
mauve-text==0.3.0
nltk==3.8.1
numpy==1.23.5
openai==0.27.4
pandas==1.5.3
python-dotenv==1.0.0
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
spacy==3.5.1
tensorflow==2.12.0
torch==2.0.0
torchdata==0.6.0
torchtext==0.15.1
tqdm==4.65.0
transformers==4.25.1